ecosystem-automation/collector-watcher/pyproject.toml
[project]
name = "collector-watcher"
version = "0.1.0"
description = "Automation tool for watching and collecting OpenTelemetry Collector component metadata"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0.1",
    "GitPython>=3.1.40",
    "PyGithub>=2.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/collector_watcher"]